+++ docs/identifier_provisioning/CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to this gear are documented in this file.
 
+## Unreleased
+
+* Updates to read in files with `utf-8-sig` to handle BOM encoding
+
 ## 1.2.4
 * Adds ErrorLogTemplate
 * Strips leading zeros from PTID in error log name