+++ docs/project_management/CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this gear are documented in this file.
 
 ## Unreleased
 
-* TBD
+* Updates to read in files with `utf-8-sig` to handle BOM encoding
 
 ## 1.0.7